+++ zapisy/apps/grade/ticket_create/models/signing_key.py
@@ -114,3 +114,53 @@ class SigningKey(models.Model):
                 error_polls.append(poll)
 
         return valid_polls, error_polls
+
+    @staticmethod
+    def get_signing_response(user, poll: Poll, signing_request: Dict) -> Dict:
+        """Responds to signing request
+
+        Returns:
+            Dictionary with keys:
+                'status': 'ERROR'
+                'message': <error_msg>
+                'id': <signing_request_id>
+            when something goes wrong, or
+            Dictionary with keys
+                'status': 'OK'
+                'signature': <ticket_signature>
+                'id': <signing_request_id>
+        """
+        error_msg = None
+        if poll.signingkey.student_used_key(user.student):
+            error_msg = "Bilet już pobrano"
+        elif not poll.is_student_entitled_to_poll(user.student):
+            error_msg = "Nie jesteś przypisany do tej ankiety"
+
+        if error_msg:
+            return {
+                'status': 'ERROR',
+                'message': error_msg,
+                'id': signing_request['id'],
+            }
+        key = SigningKey.objects.get(poll=poll)
+        signed_ticket = key.sign_ticket(signing_request['ticket'])
+        return {
+            'status': 'OK',
+            'id': signing_request['id'],
+            'signature': str(signed_ticket),
+        }
+
+    @staticmethod
+    def match_signing_requests_with_polls(signing_requests, user):
+        """For each signing request, matches it with poll corresponding to provided id, ensuring
+        that it will match only the polls user is allowed to vote in.
+
+        Raises:
+            KeyError: When there is not match between signing request and students polls.
+        """
+        matched_requests = []
+        student_polls = Poll.get_all_polls_for_student_as_dict(user.student)
+        for req in signing_requests:
+            poll = student_polls[req['id']]
+            matched_requests.append((req, poll))
+        return matched_requests
